server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
python-multipart>=0.0.9

google-genai>=1.0.0
weaviate-client>=4.9.0
PyMuPDF>=1.24.0

cachetools>=5.3.0
tenacity>=8.2.0
